pyproject.toml
[project]
name = "realmate-challenge"
version = "0.1.0"
description = "API for syncing WhatsApp events via webhooks"
authors = [{ name = "Wellington Ramos", email = "wellingtonjsramos@hotmail.com" }]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = ["django (>=5.1.6,<6.0.0)", "djangorestframework (>=3.16.0,<4.0.0)", "drf-spectacular (>=0.28.0,<0.29.0)", "djlint (>=1.36.4,<2.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pylint-django = "^2.6.1"

